[Refactor] 디렉토리 변경에 따른 import 위치 변경
